+++ HLSpriteKit/HLToolbarNode.m
@@ -504,11 +504,11 @@ static const NSTimeInterval HLToolbarSlideDuration = 0.15f;
 #endif
 }
 
-- (BOOL)addToGesture:(HLGestureRecognizer *)gestureRecognizer firstLocation:(CGPoint)sceneLocation isInside:(BOOL *)isInside
+- (BOOL)addToGesture:(HLGestureRecognizer *)gestureRecognizer firstLocation:(CGPoint)sceneLocation didAbsorbGesture:(BOOL *)didAbsorbGesture
 {
-  *isInside = YES;
 #if TARGET_OS_IPHONE
   if ([gestureRecognizer isKindOfClass:[UITapGestureRecognizer class]]) {
+    *didAbsorbGesture = YES;
     // note: Require only one tap and one touch, same as our gesture recognizer returned
     // from addsToGestureRecognizers?  I think it's okay to be non-strict.
     [gestureRecognizer addTarget:self action:@selector(handleTap:)];
@@ -516,10 +516,17 @@ static const NSTimeInterval HLToolbarSlideDuration = 0.15f;
   }
 #else
   if ([gestureRecognizer isKindOfClass:[NSClickGestureRecognizer class]]) {
+    *didAbsorbGesture = YES;
     [gestureRecognizer addTarget:self action:@selector(handleClick:)];
     return YES;
   }
 #endif
+  // note: Absorb only tap gestures.  I can easily imagine other desirable configurations
+  // but those will require a custom gesture target implementation.  This is the behavior
+  // of this gesture target since 4/2024; previously, the implementation of the gesture
+  // controller in `HLScene` would cause taps and long-presses on this gesture target to
+  // be absorbed and all other gestures to fall through.
+  *didAbsorbGesture = NO;
   return NO;
 }
 
@@ -828,10 +835,10 @@ static const NSTimeInterval HLToolbarSlideDuration = 0.15f;
             [[UIPanGestureRecognizer alloc] init] ];
 }
 
-- (BOOL)addToGesture:(HLGestureRecognizer *)gestureRecognizer firstLocation:(CGPoint)sceneLocation isInside:(BOOL *)isInside
+- (BOOL)addToGesture:(HLGestureRecognizer *)gestureRecognizer firstLocation:(CGPoint)sceneLocation didAbsorbGesture:(BOOL *)didAbsorbGesture
 {
-  *isInside = YES;
   if ([gestureRecognizer isKindOfClass:[UITapGestureRecognizer class]]) {
+    *didAbsorbGesture = YES;
     // note: Require only one touch, same as our gesture recognizer returned from
     // addsToGestureRecognizers?  I think it's okay to be non-strict.
     NSUInteger numberOfTapsRequired = [(UITapGestureRecognizer *)gestureRecognizer numberOfTapsRequired];
@@ -846,12 +853,20 @@ static const NSTimeInterval HLToolbarSlideDuration = 0.15f;
         break;
     }
   } else if ([gestureRecognizer isKindOfClass:[UILongPressGestureRecognizer class]]) {
+    *didAbsorbGesture = YES;
     [gestureRecognizer addTarget:self action:@selector(handleLongPress:)];
     return YES;
   } else if ([gestureRecognizer isKindOfClass:[UIPanGestureRecognizer class]]) {
+    *didAbsorbGesture = YES;
     [gestureRecognizer addTarget:self action:@selector(handlePan:)];
     return YES;
   }
+  // note: Absorb only gestures that are handled.  I can easily imagine other desirable
+  // configurations but those will require a different gesture target implementation.
+  // This is the behavior of this gesture target since 4/2024; previously, the
+  // implementation of the gesture controller in `HLScene` would cause swipes and pinches
+  // to fall through.
+  *didAbsorbGesture = NO;
   return NO;
 }